refactor: renomeia pasta UINodes para UINodeEditor para melhor clareza
